few_shot/requirements.txt
yacs~=0.1.8
scikit-learn
timm~=0.4.12
numpy~=1.21.0
sharedmem
git+https://github.com/openai/CLIP.git
torch~=1.7.0
PyYAML~=5.4.1
Pillow~=9.0.1
torchvision~=0.8.0
vision-evaluation>=0.2.2
vision-datasets>=0.2.0
tqdm~=4.62.3
transformers~=4.11.3
